+++ test/Functions/LoadConfigFromGlobTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\Config\Functions;
 
 use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 
 use function Shlinkio\Shlink\Config\loadConfigFromGlob;
@@ -22,8 +23,18 @@ class LoadConfigFromGlobTest extends TestCase
             ],
             'bar' => [
                 'number' => 123,
-                'array' => [1, 2, 3]
+                'array' => [1, 2, 3],
             ],
         ], $result);
+    }
+
+    #[Test]
+    #[TestWith([__DIR__ . '/../../test-resources/configs/does-not-exist.php'])]
+    #[TestWith([__DIR__ . '/../../composer.json'])]
+    #[TestWith(['does-not-exist.php'])]
+    public function nonPhpFilesAreNonExistingFilesAreSkipped(string $globPattern): void
+    {
+        $result = loadConfigFromGlob($globPattern);
+        self::assertEmpty($result);
     }
 }